ゲーム開始時のホログラム中のカメラ処理追加、SceneTitleの演出追加とテクスチャはfディスコ、SceneLoadingのテクスチャ変更、SceneGameのジッターを抑えるのとLoading明けのスクロールを消す、Player.hの変更点はホログラムのゲッター追加、Player.cppはタイトルでの処理変更とホログラム中のカメラ処理追加

+++ Source/LightManager.h
@@ -22,7 +22,7 @@ public:
 	void Activate(ID3D11DeviceContext* immediate_context);
 	void CompositActivate(ID3D11DeviceContext* immediate_context);
 	
-	const int GetNumDirLight() { return lightParam.size(); };
+	const int GetNumDirLight() { return static_cast<int>(lightParam.size()); };
 
 	void Update(float elapsedTime);
